requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
opencv-python==4.9.0.80
torch==2.2.0
torchvision==0.17.0
nest_asyncio==1.6.0
pydantic==2.7.1
pyngrok==7.1.5
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0
pytesseract